requirements.txt
fastmcp>=0.1.0
python-dotenv>=1.0.0
fastapi>=0.100.0
uvicorn>=0.22.0
python-multipart>=0.0.6
aiohttp>=3.8.0